postgres/client/requirements.txt
fastapi
uvicorn
pydantic[dotenv]
psycopg2-binary